034: Adaptive Reasoning
When the compiler says drily, "This is unlikely to be what you want,"
you realize you have found someone you can talk to.

Your only other question is, Can he prescribe meds, or is he not that
kind of doctor?

From the ML, the perils of adapting to Tuple0 aka Unit value.