Massive Simplification
As suggested by AP and implemented by APM, just use a Printer with
a "prompter" (instead of "labeller").

The core of the puzzler is that curly braces are not like parens here,
so prompt is evaluated differently than expected.  The "prompted"
(instead of "labelled") flag is partly a ruse, partly a hint.

Also modified from the suggestion, put the supposedly unprompted form
first, so the choices are, "as expected, no prompt and prompt",
"prompts both times? that has to be wrong!", and "aha! it's prompt and
then no prompt because that prompt thing is evaluated once the first time,
and the second time it doesn't matter if you set the flag.  The very fact
that it takes three or four lines to justify this answer means it's
probably wrong, but I'm sticking with it anyway."  The last, lame
filler answer is DNC for beginners who don't understand =>A syntax. Let
alone =>Unit.