Updating 'Market Precedence' to use implicits with type ascriptions
The explanation of the previous version of the puzzler was incorrect; the actual reason for the observed behaviour was the same as for 'Stringy Conversions'.

The new version behaves slightly differently, but still illustrates the potentially unexpected difference between value definition in a block vs. in a template.

See also https://github.com/scalapuzzlers/scalapuzzlers.github.com/pull/85#discussion_r76911381